Feature toggle sitemaps.
