+++ api-metadata/src/integrationTest/groovy/org/cedar/onestop/api/metadata/ETLIntegrationTests.groovy
@@ -40,9 +40,6 @@ class ETLIntegrationTests extends Specification {
   @Value('${elasticsearch.index.prefix:}${elasticsearch.index.search.flattened-granule.name}')
   private String FLAT_GRANULE_SEARCH_INDEX
 
-  @Value('${elasticsearch.index.prefix:}${elasticsearch.index.sitemap.name}')
-  String SITEMAP_INDEX
-
   @Value('${elasticsearch.index.prefix:}')
   String PREFIX
 
@@ -89,63 +86,6 @@ class ETLIntegrationTests extends Specification {
     and:
     // No flattened granules were made
     !indexed[FLAT_GRANULE_TYPE]
-  }
-
-  def 'updating sitemap with collections'() {
-    setup:
-    insertMetadataFromPath('data/COOPS/C1.xml')
-    insertMetadataFromPath('data/DEM/1.xml')
-
-    when:
-    etlService.updateSearchIndices()
-    etlService.updateSitemap()
-
-    refreshIndices()
-
-    then:
-    def indexedCols = documentsByType(COLLECTION_SEARCH_INDEX, GRANULE_SEARCH_INDEX, FLAT_GRANULE_SEARCH_INDEX)
-    def collections = indexedCols[COLLECTION_TYPE]
-    collections.size == 2
-    def indexed = searchSitemap()
-    indexed.data.size == 1
-    def sitemap = getById(SITEMAP_INDEX, indexed.data[0].id)
-    sitemap.data[0].attributes.content.size == 2
-
-    and:
-    def collectionIds = collections.collect({data -> data._id}) as Set
-    def sitemapCollectionIds = sitemap.data[0].attributes.content as Set
-    sitemapCollectionIds == collectionIds
-  }
-
-  def 'sitemap with multiple submaps'() {
-    setup:
-    insertMetadataFromPath('data/COOPS/C1.xml')
-    insertMetadataFromPath('data/DEM/1.xml')
-    insertMetadataFromPath('data/DEM/2.xml')
-    insertMetadataFromPath('data/DEM/3.xml')
-    insertMetadataFromPath('data/GHRSST/1.xml')
-    insertMetadataFromPath('data/GHRSST/2.xml')
-    insertMetadataFromPath('data/GHRSST/3.xml')
-
-    when:
-    etlService.updateSearchIndices()
-    etlService.updateSitemap()
-
-    refreshIndices()
-
-    then:
-    def indexed = searchSitemap()
-    indexed.data.size == 2
-    def submap1 = getById(SITEMAP_INDEX, indexed.data[0].id)
-    def submap2 = getById(SITEMAP_INDEX, indexed.data[1].id)
-
-    and:
-    def submapSizes = ([submap1.data[0].attributes.content.size] as Set) + ([submap2.data[0].attributes.content.size] as Set)
-    submapSizes == [6,1] as Set
-
-    and:
-    def submapIds = (submap1.data[0].attributes.content as Set) + (submap2.data[0].attributes.content as Set)
-    submapIds.size() == 7
   }
 
   def 'updating an orphan granule indexes nothing'() {
@@ -363,46 +303,6 @@ class ETLIntegrationTests extends Specification {
     ]
     def response = elasticsearchService.performRequest('GET', endpoint, request)
     return response.hits.hits.collectEntries { [(it._source.fileIdentifier): it._version] }
-  }
-
-  private Map searchSitemap() {
-    def requestBody = [
-      _source: ["lastUpdatedDate",]
-    ]
-    String searchEndpoint = "${SITEMAP_INDEX}/_search"
-    def searchResponse = elasticsearchService.performRequest('GET', searchEndpoint, requestBody )
-
-    def result = [
-      data: searchResponse.hits.hits.collect {
-        [id: it._id, type: TYPE, attributes: it._source]
-      },
-      meta: [
-          took : searchResponse.took,
-          total: searchResponse.hits.total
-      ]
-    ]
-    return result
-  }
-
-  private Map getById(String index, String id) {
-    String endpoint = "/${PREFIX}${index}/${TYPE}/${id}"
-    def response = elasticsearchService.performRequest('GET', endpoint)
-    if (response.found) {
-      return [
-          data: [[
-                     id        : response._id,
-                     type      : TYPE,
-                     attributes: response._source
-                 ]]
-      ]
-    }
-    else {
-      return [
-          status: HttpStatus.NOT_FOUND.value(),
-          title : 'No such document',
-          detail: "Record type $type with Elasticsearch ID [ ${id} ] does not exist."
-      ]
-    }
   }
 
   private refreshIndices() {
