+++ api-metadata/src/integrationTest/resources/application-integration.yml
@@ -3,3 +3,5 @@ etl:
   sitemap:
     scroll-size: 2
     collections-per-submap: 5
+
+features.sitemap: true
