+++ api-metadata/src/main/groovy/org.cedar.onestop.api.metadata/controller/AdminController.groovy
@@ -2,6 +2,7 @@ package org.cedar.onestop.api.metadata.controller
 
 import groovy.util.logging.Slf4j
 import org.cedar.onestop.api.metadata.service.ETLService
+import org.cedar.onestop.api.metadata.service.SitemapETLService
 import org.cedar.onestop.api.metadata.service.ElasticsearchService
 import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.web.bind.annotation.RequestMapping
@@ -15,24 +16,32 @@ import static org.springframework.web.bind.annotation.RequestMethod.PUT
 @RestController
 class AdminController {
 
+  @Autowired
   private ETLService etlService
-  private ElasticsearchService elasticsearchService
+
+  @Autowired(required = false)
+  private SitemapETLService sitemapEtlService
 
   @Autowired
-  AdminController(ETLService etlService, ElasticsearchService elasticsearchService) {
-    this.etlService = etlService
-    this.elasticsearchService = elasticsearchService
-  }
+  private ElasticsearchService elasticsearchService
+
+  AdminController() {}
 
   @RequestMapping(path = '/admin/index/search/rebuild', method = [GET, PUT], produces = 'application/json')
   Map rebuildSearchIndex() {
     etlService.rebuildSearchIndicesAsync()
+    if(sitemapEtlService) {
+      sitemapEtlService.rebuildSearchIndicesAsync()
+    }
     return [acknowledged: true]
   }
 
   @RequestMapping(path = '/admin/index/search/update', method = [GET, PUT], produces = 'application/json')
   Map updateSearchIndex() {
     etlService.updateSearchIndicesAsync()
+    if(sitemapEtlService) {
+      sitemapEtlService.updateSearchIndicesAsync()
+    }
     return [acknowledged: true]
   }
