+++ api-metadata/src/main/groovy/org.cedar.onestop.api.metadata/service/ElasticsearchService.groovy
@@ -31,6 +31,9 @@ class ElasticsearchService {
   @Value('${elasticsearch.index.prefix:}${elasticsearch.index.search.flattened-granule.name}')
   private String FLAT_GRANULE_SEARCH_INDEX
 
+  @Value('${features.sitemap}')
+  private String SITEMAP_ENABLED
+
   @Value('${elasticsearch.index.prefix:}${elasticsearch.index.sitemap.name}')
   private String SITEMAP_INDEX
 
@@ -65,7 +68,9 @@ class ElasticsearchService {
     ensureIndex(COLLECTION_SEARCH_INDEX)
     ensureIndex(GRANULE_SEARCH_INDEX)
     ensureIndex(FLAT_GRANULE_SEARCH_INDEX)
-    ensureIndex(SITEMAP_INDEX)
+    if(SITEMAP_ENABLED) {
+      ensureIndex(SITEMAP_INDEX)
+    }
   }
 
   public void ensurePipelines() {
@@ -118,7 +123,9 @@ class ElasticsearchService {
     drop(COLLECTION_SEARCH_INDEX)
     drop(GRANULE_SEARCH_INDEX)
     drop(FLAT_GRANULE_SEARCH_INDEX)
-    drop(SITEMAP_INDEX)
+    if(SITEMAP_ENABLED) {
+      drop(SITEMAP_INDEX)
+    }
   }
 
   public void drop(String indexName) {
